Testimony Search Fleshed Out
