+++ components/search/SortBy.tsx
@@ -3,32 +3,6 @@ import { SortByItem } from "instantsearch.js/es/connectors/sort-by/connectSortBy
 import Select from "react-select"
 import styled from "styled-components"
 
-const items: SortByItem[] = [
-  /**
-   * uncomment the commented sections below once Prod is pointing at the correct data
-   */
-
-  // {
-  //   label: "Sort by Most Recent Testimony",
-  //   value: "bills/sort/latestTestimonyAt:desc"
-  // },
-  {
-    label: "Sort by Relevance",
-    value: "bills/sort/_text_match:desc,testimonyCount:desc"
-  },
-  // { label: "Sort by Testimony Count", value: "bills/sort/testimonyCount:desc" },
-  {
-    label: "Sort by Cosponsor Count",
-    value: "bills/sort/cosponsorCount:desc"
-  },
-  {
-    label: "Sort by Next Hearing Date",
-    value: "bills/sort/nextHearingAt:desc"
-  }
-]
-
-export const initialSortByValue = items[0].value
-
 const StyledSelect = styled(Select)`
   .s__control {
     background-color: var(--bs-blue);
@@ -52,7 +26,7 @@ const StyledSelect = styled(Select)`
   }
 `
 
-export const SortBy = () => {
+export const SortBy = ({ items }: { items: SortByItem[] }) => {
   const sortBy = useSortBy({ items }),
     selected = items.find(i => i.value === sortBy.currentRefinement)!
   return (
