+++ components/search/bills/BillHit.tsx
@@ -10,8 +10,8 @@ import { format, fromUnixTime } from "date-fns"
 import { Hit } from "instantsearch.js"
 import Link from "next/link"
 import styled from "styled-components"
-import { Card, Col } from "../bootstrap"
-import { formatBillId } from "../formatting"
+import { Card, Col } from "../../bootstrap"
+import { formatBillId } from "../../formatting"
 
 type BillRecord = {
   number: string
