+++ components/search/bills/BillSearch.tsx
@@ -9,16 +9,17 @@ import {
 import { currentGeneralCourt } from "functions/src/shared"
 import styled from "styled-components"
 import TypesenseInstantSearchAdapter from "typesense-instantsearch-adapter"
-import { Col, Row } from "../bootstrap"
+import { Col, Row } from "../../bootstrap"
 import { BillHit } from "./BillHit"
-import { getServerConfig } from "./common"
-import { NoResults } from "./NoResults"
-import { ResultCount } from "./ResultCount"
-import { SearchContainer } from "./SearchContainer"
-import { SearchErrorBoundary } from "./SearchErrorBoundary"
-import { initialSortByValue, SortBy } from "./SortBy"
-import { useRefinements } from "./useRefinements"
-import { useRouting } from "./useRouting"
+import { getServerConfig } from "../common"
+import { NoResults } from "../NoResults"
+import { ResultCount } from "../ResultCount"
+import { SearchContainer } from "../SearchContainer"
+import { SearchErrorBoundary } from "../SearchErrorBoundary"
+import { SortBy } from "../SortBy"
+import { useRouting } from "../useRouting"
+import { useBillRefinements } from "./useBillRefinements"
+import { SortByItem } from "instantsearch.js/es/connectors/sort-by/connectSortBy"
 
 const searchClient = new TypesenseInstantSearchAdapter({
   server: getServerConfig(),
@@ -27,6 +28,32 @@ const searchClient = new TypesenseInstantSearchAdapter({
     exclude_fields: "body"
   }
 }).searchClient
+
+const items: SortByItem[] = [
+  /**
+   * uncomment the commented sections below once Prod is pointing at the correct data
+   */
+
+  // {
+  //   label: "Sort by Most Recent Testimony",
+  //   value: "bills/sort/latestTestimonyAt:desc"
+  // },
+  {
+    label: "Sort by Relevance",
+    value: "bills/sort/_text_match:desc,testimonyCount:desc"
+  },
+  // { label: "Sort by Testimony Count", value: "bills/sort/testimonyCount:desc" },
+  {
+    label: "Sort by Cosponsor Count",
+    value: "bills/sort/cosponsorCount:desc"
+  },
+  {
+    label: "Sort by Next Hearing Date",
+    value: "bills/sort/nextHearingAt:desc"
+  }
+]
+
+export const initialSortByValue = items[0].value
 
 export const BillSearch = () => (
   <SearchErrorBoundary>
@@ -66,7 +93,7 @@ const useSearchStatus = () => {
 }
 
 const Layout = () => {
-  const refinements = useRefinements()
+  const refinements = useBillRefinements()
   const status = useSearchStatus()
   return (
     <SearchContainer>
@@ -78,7 +105,7 @@ const Layout = () => {
         <Col className="d-flex flex-column">
           <RefinementRow>
             <ResultCount className="flex-grow-1 m-1" />
-            <SortBy />
+            <SortBy items={items} />
             {refinements.show}
           </RefinementRow>
           <CurrentRefinements className="mt-2 mb-2" />
