+++ components/search/index.ts
@@ -1,2 +1,2 @@
-export * from "./BillSearch"
-export * from "./service"
+export * from "./bills/BillSearch"
+export * from "./bills/service"
