+++ components/search/useRefinements.tsx
@@ -5,8 +5,6 @@ import {
 } from "@alexjball/react-instantsearch-hooks-web"
 import { faFilter } from "@fortawesome/free-solid-svg-icons"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { generalCourts } from "functions/src/shared"
-import { RefinementListItem } from "instantsearch.js/es/connectors/refinement-list/connectRefinementList"
 import { useCallback, useEffect, useState } from "react"
 import styled from "styled-components"
 import { useMediaQuery } from "usehooks-ts"
@@ -27,7 +25,11 @@ const useHasRefinements = () => {
   return refinements.length !== 0
 }
 
-export const useRefinements = () => {
+export const useRefinements = ({
+  refinementProps
+}: {
+  refinementProps: any[]
+}) => {
   const inline = useMediaQuery("(min-width: 768px)")
   const [show, setShow] = useState(false)
   const handleClose = useCallback(() => setShow(false), [])
@@ -36,45 +38,6 @@ export const useRefinements = () => {
   useEffect(() => {
     if (inline) setShow(false)
   }, [inline])
-
-  const baseProps = { limit: 500, searchable: true }
-  const refinementProps = [
-    useRefinementListUiProps({
-      transformItems: useCallback(
-        (i: RefinementListItem[]) =>
-          i
-            .map(i => ({
-              ...i,
-              label: generalCourts[i.value as any]?.Name ?? i.label
-            }))
-            .sort((a, b) => Number(b.value) - Number(a.value)),
-        []
-      ),
-      attribute: "court",
-      searchablePlaceholder: "General Court",
-      ...baseProps
-    }),
-    useRefinementListUiProps({
-      attribute: "currentCommittee",
-      ...baseProps,
-      searchablePlaceholder: "Current Committee"
-    }),
-    useRefinementListUiProps({
-      attribute: "city",
-      searchablePlaceholder: "City",
-      ...baseProps
-    }),
-    useRefinementListUiProps({
-      attribute: "primarySponsor",
-      ...baseProps,
-      searchablePlaceholder: "Primary Sponsor"
-    }),
-    useRefinementListUiProps({
-      attribute: "cosponsors",
-      ...baseProps,
-      searchablePlaceholder: "Cosponsor"
-    })
-  ]
 
   const refinements = (
     <>
